Grid_Search__
In this, from sklearn module imported datasets. From which I get iris data. Then I used pandas to learn the datasets. Then I used logistic regression to fit the dataset and create my model. Logistic Regression has one of it's parameter that is C which is complexity so we created different values of C for learning our model performances when changing the value of C. However, the default of value in a logistic regression model is 1. But we capture different results with changing C values. After doing some code we find that the best result we get of when value of C is 1.75. Although the model performs well, we don't split our dataset into training and testing of data. Next time I also apply testing data results and then change the value of C and then examine the e results. Also, change the algorithm from logistic regression to another model for flexibility of learning grid search in more detail and it's working.